requirements.txt
markdownify~=1.1.0
smolagents~=1.14.0
requests~=2.32.3
duckduckgo_search~=8.0.1
pandas
requests
pytz~=2025.2
PyYAML~=6.0.2
openai>=1.0.0